Improved route based translation
