+++ src/lib/types.d.ts
@@ -1,5 +1,3 @@
-import type { SvelteComponentTyped } from "svelte";
-
 export interface Page {
 	slug: string;
 	index: number;
@@ -13,11 +11,3 @@ export interface Category {
 	name: string;
 	pages: Page[];
 }
-
-export interface Post {
-	default: Promise<SvelteComponentTyped>;
-}
-
-export type PostResolver = () => Promise<Post | PostResolver>;
-
-export type GlobImport = Record<string, PostResolver>;
