+++ src/routes/[[locale]]/+page.svelte
